backend/requirements.txt
# Production Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Database
motor==3.3.2  # Async MongoDB
pymongo==4.6.1
redis==5.0.1
aioredis==2.0.1

# Message Queue
aiokafka==0.10.0
kafka-python==2.0.2

# ML Integration
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
onnxruntime==1.16.3
joblib==1.3.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
python-multipart==0.0.6

# Monitoring and Logging
prometheus-client==0.19.0
structlog==24.1.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
orjson==3.9.10  # Fast JSON serialization

# Rate Limiting
slowapi==0.1.9

# Distributed Tracing
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-jaeger==1.21.0

# Async utilities
asyncio==3.4.3
